% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{CalculationFunctions}
\alias{CalculationFunctions}
\alias{getRuns}
\alias{doParCalc}
\alias{calcPCC}
\alias{calcPCCPair}
\alias{calcrPCCPair}
\alias{calcHyperPPair}
\alias{calcrHyperPPair}
\alias{calcOvPair}
\alias{calcJCPair}
\alias{calcrJCPair}
\alias{calcOccDiff}
\alias{calcOccDiffPair}
\title{Metric calculation functions.}
\usage{
getRuns(x, y)

doParCalc(z, FUN)

calcPCC(z)

calcPCCPair(x, y)

calcrPCCPair(x, y)

calcHyperPPair(x, y)

calcrHyperPPair(x, y)

calcOvPair(x, y)

calcJCPair(x, y)

calcrJCPair(x, y)

calcOccDiff(z)

calcOccDiffPair(x, y)
}
\arguments{
\item{x}{A single column from the presence-absence input matrix to compare
to \code{y}.}

\item{y}{The second single column with which to compare to \code{x}.}

\item{z}{The presence-absence input matrix.}

\item{FUN}{The function perform the matrix-wide all-against-all
calculations, optionally in parallel.}
}
\value{
For \code{\link[=getRuns]{getRuns()}}: a list of \code{x} and \code{y}, runs-adjusted.

For all other functions: either a single value if comparing two columns,
or a matrix of values if comparing all-by-all.
}
\description{
Comparison functions to calculate the various metrics on the presence-absence
input matrix.
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
