% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{phylocorrpkg-pkg}
\alias{phylocorrpkg-pkg}
\title{phylocorrpkg: Phylogenetic correlations across large trees}
\description{
Functional predictions from correlations across large phylogenetic
trees using presence-absence data.
}
