% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{ProbabilityFunctions}
\alias{ProbabilityFunctions}
\alias{mergeScoresAndAnnotations}
\alias{getProbMatchFunSingle}
\alias{getProbMatchDfDouble}
\alias{getProbMatchMatrixDouble}
\alias{calcMatchingProbsSingle}
\alias{calcMatchingProbsDouble}
\title{Probability prediction functions.}
\usage{
mergeScoresAndAnnotations(scores, annotations, highScoreIsBest = TRUE)

getProbMatchFunSingle(
  merged,
  highScoreIsBest = TRUE,
  maxProb = 1,
  useMeanSmoothing = FALSE,
  windowSize = max(1, round(1e-04 * nrow(merged)))
)

getProbMatchDfDouble(
  df1,
  df2,
  len1 = 21,
  len2 = 21,
  lims1 = range(df1$Score),
  lims2 = range(df2$Score),
  bw1Match = MASS::bandwidth.nrd(df1$Score[df1$MatchingAnno]) * 2,
  bw1NoMatch = MASS::bandwidth.nrd(df1$Score[!df1$MatchingAnno]) * 2,
  bw2Match = MASS::bandwidth.nrd(df2$Score[df2$MatchingAnno]) * 2,
  bw2NoMatch = MASS::bandwidth.nrd(df2$Score[!df2$MatchingAnno]) * 2
)

getProbMatchMatrixDouble(
  CombinedDf,
  useMeanSmoothing = FALSE,
  windowSize1 = 3,
  windowSize2 = 3
)

calcMatchingProbsSingle(scores, predFUN)

calcMatchingProbsDouble(scores1, scores2, predMatrix)
}
\arguments{
\item{scores}{A square matrix of pair-wise scores.}

\item{annotations}{A list of annotations, such as that generated by
\code{\link[=cleanAnnotations]{cleanAnnotations()}}.}

\item{highScoreIsBest}{Whether the score is one where higher values represent
better scores. An example of when this is true would be the Jaccard
coefficient, and false would be P-values (though it is strongly recommended
to \code{-log10()} transform P-values in order to obtain a linear range of scores).}

\item{merged}{The output of \code{\link[=mergeScoresAndAnnotations]{mergeScoresAndAnnotations()}}.}

\item{maxProb}{What is the max possible predicted probability? If the
available scores do not predict anywhere near 100\%, it may be worthwhile
manually lower this to a more sensible value as to not generate
overly-optimistic predictions.}

\item{useMeanSmoothing}{Whether to smooth the probability function/matrix
using moving-window mean smoothing.}

\item{windowSize}{The window size for moving-window mean smoothing.}

\item{df1}{The first set of scores, as generated by \code{\link[=mergeScoresAndAnnotations]{mergeScoresAndAnnotations()}}.}

\item{df2}{The second set of scores, as generated by \code{\link[=mergeScoresAndAnnotations]{mergeScoresAndAnnotations()}}.}

\item{len1}{The number of intervals of use in the final prediction matrix,
corresponding to the first set of scores.}

\item{len2}{The number of intervals of use in the final prediction matrix,
corresponding to the second set of scores.}

\item{lims1}{The min and max values possible for the first set of scores.
By default, this is detected from the min and max values of the input,
but it can be manually set if the input data does not represent the full
possible range.}

\item{lims2}{The min and max values possible for the second set of scores.
By default, this is detected from the min and max values of the input,
but it can be manually set if the input data does not represent the full
possible range.}

\item{bw1Match}{The bandwidth value used by \code{\link[KernSmooth:bkde2D]{KernSmooth::bkde2D()}} for the
subset of the first set of scores which represent an annotation match.}

\item{bw1NoMatch}{The bandwidth value used by \code{\link[KernSmooth:bkde2D]{KernSmooth::bkde2D()}} for the
subset of the first set of scores which do not represent an annotation match.}

\item{bw2Match}{The bandwidth value used by \code{\link[KernSmooth:bkde2D]{KernSmooth::bkde2D()}} for the
subset of the second set of scores which represent an annotation match.}

\item{bw2NoMatch}{The bandwidth value used by \code{\link[KernSmooth:bkde2D]{KernSmooth::bkde2D()}} for the
subset of the second set of scores which do not represent an annotation match.}

\item{CombinedDf}{The out of \code{\link[=getProbMatchDfDouble]{getProbMatchDfDouble()}}.}

\item{windowSize1}{The window size for moving-window mean smoothing, for the
first set of scores.}

\item{windowSize2}{The window size for moving-window mean smoothing, for the
second set of scores.}

\item{predFUN}{The prediction function as generated by \code{\link[=getProbMatchFunSingle]{getProbMatchFunSingle()}}.}

\item{scores1}{The first set of scores to use for prediction.}

\item{scores2}{The second set of scores to use for prediction.}

\item{predMatrix}{The prediction matrix as generated by
\code{\link[=getProbMatchMatrixDouble]{getProbMatchMatrixDouble()}}.}
}
\value{
For \code{\link[=mergeScoresAndAnnotations]{mergeScoresAndAnnotations()}}: The input score matrix is converted to
\code{data.frame} with pair-wise scores, keeping only families which have
associated annotation. An additional logical column indicates the presence
of a matching annotation between the pairs.

For \code{\link[=getProbMatchFunSingle]{getProbMatchFunSingle()}}: A function which can be used to predict
the probability of a matching annotation/function using the input scoring
method.

For \code{\link[=getProbMatchDfDouble]{getProbMatchDfDouble()}}: A \code{data.frame} with \code{x} and \code{y} coordinates
for the annotation/function prediction matrix, which associated
probability data.

For \code{\link[=getProbMatchMatrixDouble]{getProbMatchMatrixDouble()}}: A matrix of probabilities which can be
used to predict the probability of a annotation/function match based on
two input scores.

For \code{\link[=calcMatchingProbsSingle]{calcMatchingProbsSingle()}}, \code{\link[=calcMatchingProbsDouble]{calcMatchingProbsDouble()}}: A matrix of
probabilities of matching annotation/function based on the input score.
}
\description{
Functions to be used to generate the prediction functions/matrices.
}
\author{
Benjamin Jean-Marie Tremblay, \email{b2tremblay@uwaterloo.ca}
}
